refactor: move theme to uad-gui; drop core gui deps; mark PmCommand must-use
