+++ crates/uad-core/src/lib.rs
@@ -11,7 +11,6 @@ pub mod adb;
 pub mod config;
 pub mod save;
 pub mod sync;
-pub mod theme;
 pub mod uad_lists;
 pub mod update;
 pub mod utils;
