+++ crates/uad-gui/Cargo.toml
@@ -22,7 +22,7 @@ no-self-update = []
 img = ["image", "iced/image"]
 
 [dependencies]
-uad-core = { path = "../uad-core", features = ["gui", "self-update"] }
+uad-core = { path = "../uad-core", features = ["self-update"] }
 iced.workspace = true
 image = { workspace = true, optional = true }
 rfd.workspace = true
@@ -54,7 +54,6 @@ format_push_string = "warn"
 large_include_file = "warn"
 shadow_unrelated = "warn"
 struct_field_names = "allow" # annoying
-module_name_repetitions = "allow" # annoying
 
 disallowed_types = "deny"
 disallowed_methods = "deny"
