+++ crates/uad-gui/src/theme.rs
@@ -1,72 +1,219 @@
 use iced::theme::{self, Mode, Palette, Style};
+use iced::{Color, color};
+use std::sync::LazyLock;
 
-pub use uad_core::theme::{BaseColors, BrightColors, ColorPalette, NormalColors, OS_COLOR_SCHEME};
+/*
+In-memory caching.
+This fixes the perf bug
+caused by Iced repeatedly calling `palette`.
 
-/// GUI-local wrapper around the core Theme to satisfy orphan rules for
-/// iced's Catalog traits.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct Theme(pub uad_core::theme::Theme);
+Coincidentally, this also ensures consistent colors across the GUI,
+at the cost of requiring a restart to update the palette.
+(this is just a patch, not a fix)
+*/
+pub static OS_COLOR_SCHEME: LazyLock<dark_light::Mode> =
+    LazyLock::new(|| dark_light::detect().unwrap_or(dark_light::Mode::Unspecified));
 
-#[allow(
-    non_upper_case_globals,
-    reason = "Keep variant-like names matching core Theme"
-)]
+#[derive(Default, Debug, PartialEq, Eq, Copy, Clone)]
+/// Color scheme
+pub enum Theme {
+    #[default]
+    /// `Dark` or `Light`, according to `dark_light`
+    Auto,
+    /// `Dark`-ish and purple
+    Lupin,
+    /// white on black
+    Dark,
+    /// black on white
+    Light,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub struct BaseColors {
+    pub background: Color,
+    pub foreground: Color,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub struct NormalColors {
+    pub primary: Color,
+    #[allow(dead_code, reason = "Reserved for future palette updates")]
+    pub secondary: Color,
+    pub surface: Color,
+    pub error: Color,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub struct BrightColors {
+    pub primary: Color,
+    pub secondary: Color,
+    pub surface: Color,
+    pub error: Color,
+}
+
+#[derive(Debug, Clone, Copy)]
+pub struct ColorPalette {
+    pub base: BaseColors,
+    pub normal: NormalColors,
+    pub bright: BrightColors,
+}
+
 impl Theme {
-    pub const Auto: Self = Self(uad_core::theme::Theme::Auto);
-    pub const Lupin: Self = Self(uad_core::theme::Theme::Lupin);
-    pub const Dark: Self = Self(uad_core::theme::Theme::Dark);
-    pub const Light: Self = Self(uad_core::theme::Theme::Light);
-
     pub const ALL: [Self; 4] = [Self::Auto, Self::Lupin, Self::Dark, Self::Light];
 
+    #[allow(
+        clippy::unreadable_literal,
+        reason = "https://github.com/Universal-Debloater-Alliance/universal-android-debloater-next-generation/pull/578#discussion_r1759653408"
+    )]
+    /// This `fn` _could_ be `const`,
+    /// but `deref`ing a lazy-`static` is non-`const`.
     #[must_use]
     pub fn palette(self) -> ColorPalette {
-        self.0.palette()
+        const DARK: ColorPalette = ColorPalette {
+            base: BaseColors {
+                background: color!(0x111111),
+                foreground: color!(0x1C1C1C),
+            },
+            normal: NormalColors {
+                primary: color!(0x5E4266),
+                secondary: color!(0x386E50),
+                surface: color!(0x828282),
+                error: color!(0x992B2B),
+            },
+            bright: BrightColors {
+                primary: color!(0xBA84FC),
+                secondary: color!(0x49EB7A),
+                surface: color!(0xE0E0E0),
+                error: color!(0xC13047),
+            },
+        };
+        const LIGHT: ColorPalette = ColorPalette {
+            base: BaseColors {
+                background: color!(0xEEEEEE),
+                foreground: color!(0xE0E0E0),
+            },
+            normal: NormalColors {
+                primary: color!(0x818181),
+                secondary: color!(0xF9D659),
+                surface: color!(0x818181),
+                error: color!(0x992B2B),
+            },
+            bright: BrightColors {
+                primary: color!(0x673AB7),
+                secondary: color!(0x3797A4),
+                surface: color!(0x000000),
+                error: color!(0xC13047),
+            },
+        };
+        const LUPIN: ColorPalette = ColorPalette {
+            base: BaseColors {
+                background: color!(0x282A36),
+                foreground: color!(0x353746),
+            },
+            normal: NormalColors {
+                primary: color!(0x58406F),
+                secondary: color!(0x386E50),
+                surface: color!(0xA2A4A3),
+                error: color!(0xA13034),
+            },
+            bright: BrightColors {
+                primary: color!(0xBD94F9),
+                secondary: color!(0x49EB7A),
+                surface: color!(0xF4F8F3),
+                error: color!(0xE63E6D),
+            },
+        };
+        match self {
+            Self::Dark => DARK,
+            Self::Light => LIGHT,
+            Self::Lupin => LUPIN,
+            Self::Auto => match *OS_COLOR_SCHEME {
+                dark_light::Mode::Light => LIGHT,
+                dark_light::Mode::Dark | dark_light::Mode::Unspecified => DARK,
+            },
+        }
     }
 }
 
-impl From<uad_core::theme::Theme> for Theme {
-    fn from(value: uad_core::theme::Theme) -> Self {
-        Self(value)
-    }
-}
-
-impl From<Theme> for uad_core::theme::Theme {
-    fn from(value: Theme) -> Self {
-        value.0
+impl std::fmt::Display for Theme {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                Self::Dark => "Dark",
+                Self::Light => "Light",
+                Self::Lupin => "Lupin",
+                Self::Auto => "Auto (follow system theme)",
+            }
+        )
     }
 }
 
 /// Converts a string to the GUI's Theme type
 #[must_use]
 pub fn string_to_theme(theme: &str) -> Theme {
-    Theme(uad_core::theme::string_to_theme(theme))
+    match theme {
+        "Lupin" => Theme::Lupin,
+        "Dark" => Theme::Dark,
+        "Light" => Theme::Light,
+        _ => Theme::Auto,
+    }
 }
 
 impl theme::Base for Theme {
     fn default(preference: Mode) -> Self {
-        Self(<uad_core::theme::Theme as theme::Base>::default(preference))
+        match preference {
+            Mode::Light => Self::Light,
+            Mode::Dark => Self::Dark,
+            Mode::None => match *OS_COLOR_SCHEME {
+                dark_light::Mode::Light => Self::Light,
+                dark_light::Mode::Dark | dark_light::Mode::Unspecified => Self::Dark,
+            },
+        }
     }
 
     fn mode(&self) -> Mode {
-        <uad_core::theme::Theme as theme::Base>::mode(&self.0)
+        let resolved = match self {
+            Self::Auto => match *OS_COLOR_SCHEME {
+                dark_light::Mode::Light => Self::Light,
+                dark_light::Mode::Dark | dark_light::Mode::Unspecified => Self::Dark,
+            },
+            other => *other,
+        };
+
+        match resolved {
+            Self::Light => Mode::Light,
+            _ => Mode::Dark,
+        }
     }
 
     fn base(&self) -> Style {
-        <uad_core::theme::Theme as theme::Base>::base(&self.0)
+        let p = Self::palette(*self);
+        Style {
+            background_color: p.base.background,
+            text_color: p.bright.surface,
+        }
     }
 
     fn palette(&self) -> Option<Palette> {
-        <uad_core::theme::Theme as theme::Base>::palette(&self.0)
+        let p = Self::palette(*self);
+        Some(Palette {
+            background: p.base.background,
+            text: p.bright.surface,
+            primary: p.normal.primary,
+            success: p.bright.secondary,
+            warning: p.normal.surface,
+            danger: p.normal.error,
+        })
     }
 
     fn name(&self) -> &str {
-        <uad_core::theme::Theme as theme::Base>::name(&self.0)
-    }
-}
-
-impl std::fmt::Display for Theme {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        std::fmt::Display::fmt(&self.0, f)
+        match self {
+            Self::Dark => "Dark",
+            Self::Light => "Light",
+            Self::Lupin => "Lupin",
+            Self::Auto => "Auto",
+        }
     }
 }
