+++ flake.nix
@@ -27,10 +27,10 @@
             with pkgs;
             [
               libglvnd
-              xorg.libX11
-              xorg.libXcursor
-              xorg.libXi
-              xorg.libXrandr
+              libX11
+              libXcursor
+              libXi
+              libXrandr
               libxkbcommon
               wayland
             ]
